backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-extensions>=3.2.0
celery>=5.4.0
django-celery-beat>=2.6.0
redis>=5.0.0
psycopg2-binary>=2.9.9
scikit-learn>=1.4.0
pandas>=2.0.0
numpy>=1.26.0
joblib>=1.3.0
Pillow>=10.0.0
httpx>=0.27.0
weasyprint>=60.0
reportlab>=4.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0
